+++ app/src/main/java/com/hello/seoulnuri/model/course/CourseStarData.kt
@@ -4,7 +4,7 @@ package com.hello.seoulnuri.model.course
  * Created by shineeseo on 2018. 9. 12..
  */
 data class CourseStarData (
-        var cour_star : Int?,
-        var cour_star_count : Int?
+        var cour_star : Double?,
+        var cour_star_count : Double?
 )
